+++ sale_order_line_date/i18n/sk.po
@@ -11,12 +11,12 @@ msgstr ""
 "POT-Creation-Date: 2018-01-27 03:53+0000\n"
 "PO-Revision-Date: 2018-01-27 03:53+0000\n"
 "Last-Translator: OCA Transbot <transbot@odoo-community.org>, 2018\n"
-"Language-Team: Turkish (Turkey) (https://www.transifex.com/oca/teams/23907/tr_TR/)\n"
+"Language-Team: Slovak (https://www.transifex.com/oca/teams/23907/sk/)\n"
 "MIME-Version: 1.0\n"
 "Content-Type: text/plain; charset=UTF-8\n"
 "Content-Transfer-Encoding: \n"
-"Language: tr_TR\n"
-"Plural-Forms: nplurals=1; plural=0;\n"
+"Language: sk\n"
+"Plural-Forms: nplurals=3; plural=(n==1) ? 0 : (n>=2 && n<=4) ? 1 : 2;\n"
 
 #. module: sale_order_line_date
 #: model:ir.model.fields,field_description:sale_order_line_date.field_sale_order_line_requested_date
@@ -26,9 +26,9 @@ msgstr ""
 #. module: sale_order_line_date
 #: model:ir.model,name:sale_order_line_date.model_sale_order
 msgid "Sales Order"
-msgstr "Satış emri"
+msgstr "Objednávka predaja"
 
 #. module: sale_order_line_date
 #: model:ir.model,name:sale_order_line_date.model_sale_order_line
 msgid "Sales Order Line"
-msgstr "Sipariş emri satırı "
+msgstr ""
