+++ sale_order_line_date/tests/test_sale_order_line_date.py
@@ -3,10 +3,11 @@
 # © 2016 AvanzOSC
 # © 2016 Serv. Tecnol. Avanzados - Pedro M. Baeza
 # © 2016 Eficent Business and IT Consulting Services, S.L.
+# Copyright 2017 Serpent Consulting Services Pvt. Ltd.
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from openerp.tests.common import TransactionCase
-from openerp import fields
+from odoo.tests.common import TransactionCase
+from odoo import fields
 import datetime
 
 
@@ -15,38 +16,23 @@ class TestSaleOrderLineDates(TransactionCase):
 
     def setUp(self):
         """Setup a Sale Order with 4 lines.
-        And prepare procurements
         """
         super(TestSaleOrderLineDates, self).setUp()
         customer = self.env.ref('base.res_partner_3')
-        self.product_ctg_model = self.env['product.category']
-        self.company = self.env.ref('base.main_company')
         price = 100.0
-        qty = 1000
-        p1 = self._create_product(price)
-        self._update_qty(p1, qty)
+        qty = 5
+        product_id = self.env.ref('product.product_product_7')
         today = datetime.datetime.now()
         dt1 = today + datetime.timedelta(days=9)
         dt2 = today + datetime.timedelta(days=10)
         self.dt3 = today + datetime.timedelta(days=3)
         self.sale1 = self._create_sale_order(customer, dt2)
-        self.sale_line1 = self._create_sale_order_line(
-            self.sale1, p1, qty, price, dt1
-        )
-        self.sale_line2 = self._create_sale_order_line(
-            self.sale1, p1, qty, price, dt2
-        )
-
-    def _create_product(self, price):
-        category = self.env.ref('product.product_category_1')
-        product = self.env['product.product'].create({
-            'name': 'test_product',
-            'type': 'product',
-            'standard_price': price,
-            'list_price': price,
-            'categ_id': category.id,
-        })
-        return product
+        self.sale_line1 = self._create_sale_order_line(self.sale1, product_id,
+                                                       qty, price, dt1)
+        self.sale_line2 = self._create_sale_order_line(self.sale1, product_id,
+                                                       qty, price, dt2)
+        self.sale_line2.write({'requested_date': dt2})
+        self.sale1.action_confirm()
 
     def _create_sale_order(self, customer, date):
         sale = self.env['sale.order'].create({
@@ -67,33 +53,11 @@ class TestSaleOrderLineDates(TransactionCase):
             'requested_date': date})
         return sale_line
 
-    def _update_qty(self, product, qty):
-        location_stock = self.env.ref('stock.stock_location_stock')
-        wiz_obj = self.env['stock.change.product.qty']
-        wiz = wiz_obj.create({'product_id': product.id,
-                              'new_quantity': qty,
-                              'location_id':  location_stock.id,
-                              })
-        wiz.change_product_qty()
-
-    def test_procurement_scheduled_date(self):
-        """True when matches the requested date in the sale_order_line"""
-        self.sale1.action_confirm()
-        procurements = self.env['procurement.order'].search([
-            ('origin', '=', self.sale1.name),
-            ('date_planned', '=', self.sale_line1.requested_date)])
-        self.assertEqual(len(procurements), 1)
-        procurements = self.env['procurement.order'].search([
-            ('origin', '=', self.sale1.name),
-            ('date_planned', '=', self.sale_line2.requested_date)])
-        self.assertEqual(len(procurements), 1)
-
     def test_on_change_requested_date(self):
         """True when the requested date in the sale_order_line
         matches the requested date in the sale order"""
         req_date = fields.Datetime.to_string(self.dt3)
         self.sale1.write({'requested_date': self.dt3})
-        result = self.sale1.onchange_requested_date(self.sale1.requested_date,
-                                                    self.sale1.commitment_date)
+        result = self.sale1.onchange_requested_date()
         for line in result['value']['order_line']:
             self.assertEqual(line[2]['requested_date'], req_date)
