+++ sale_order_line_date/views/sale_order_view.xml
@@ -1,31 +1,31 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<openerp>
-    <data>
-        <record model="ir.ui.view" id="sale_order_requested_date_form_view">
-            <field name="name">sale.order.requested.date.form</field>
-            <field name="model">sale.order</field>
-            <field name="inherit_id" ref="sale_order_dates.view_order_form_inherit_sale_stock_inherit_sale_order_dates"/>
-            <field name="arch" type="xml">
-                <field name="order_line" position="attributes">
-                    <attribute name="context">{'default_requested_date':requested_date}</attribute>
-                </field>
-                <xpath expr="//field[@name='order_line']/form//field[@name='tax_id']" position="after">
-                    <field name="requested_date" />
-                </xpath>
-                <xpath expr="//field[@name='order_line']/tree/field[@name='price_unit']" position="after">
-                    <field name="requested_date"/>
-                </xpath>
+<odoo>
+
+    <record model="ir.ui.view" id="sale_order_requested_date_form_view">
+        <field name="name">sale.order.requested.date.form</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale_order_dates.view_order_form_inherit_sale_stock_inherit_sale_order_dates"/>
+        <field name="arch" type="xml">
+            <field name="order_line" position="attributes">
+                <attribute name="context">{'default_requested_date':requested_date}</attribute>
             </field>
-        </record>
-        <record id="sale_order_line_ext_tree_view" model="ir.ui.view">
-            <field name="name">sale.order.line.ext.tree</field>
-            <field name="model">sale.order.line</field>
-            <field name="inherit_id" ref="sale.view_order_line_tree" />
-            <field name="arch" type="xml">
-                <field name="name" position="after">
-                    <field name="requested_date"/>
-                </field>
+            <xpath expr="//field[@name='order_line']/form//field[@name='tax_id']" position="after">
+                <field name="requested_date" />
+            </xpath>
+            <xpath expr="//field[@name='order_line']/tree/field[@name='price_unit']" position="after">
+                <field name="requested_date"/>
+            </xpath>
+        </field>
+    </record>
+    <record id="sale_order_line_ext_tree_view" model="ir.ui.view">
+        <field name="name">sale.order.line.ext.tree</field>
+        <field name="model">sale.order.line</field>
+        <field name="inherit_id" ref="sale.view_order_line_tree" />
+        <field name="arch" type="xml">
+            <field name="name" position="after">
+                <field name="requested_date"/>
             </field>
-        </record>
-    </data>
-</openerp>
+        </field>
+    </record>
+
+</odoo>
